Added bulk of Ice elemental debuff
Added enemy slow, thawing and shattering damage. Still need to add a way to proc shatter. That includes single-ing out one enemy and calling Shatter() on them.

+++ Assets/Scripts/SmoulderingCorpseBehaviour.cs
@@ -8,7 +8,7 @@ public class SmoulderingCorpseBehaviour : FireElementClass
     public int burnStackCount = 0;
     public float tickRate = 1f;
 
-    public int burnTimer;
+    public float burnTimer;
 
     bool allowSpreadInvoke = true; //this stops multiple Invokes from being played at the same time
